refactor(redis_ops): extract formatting module
Extract value formatting helper functions into dedicated formatting.rs module:
- escape_quotes: Escapes double quotes in strings
- quoted: Wraps strings with escaped quotes
- append_expire_command: Adds EXPIRE command if TTL exists

All 4 formatting tests moved with the functions.
